rename data processing package to dfm_data (it did not work as a package efter monorepo refactoring)
