+++ data-processing/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "dfm"
-version = "2.0.0"
+name = "dfm_data"
+version = "2.0.1"
 description = "Code for the Danish Foundation Models"
 classifiers = [
   "Environment :: Console",
@@ -22,7 +22,7 @@ dependencies = [
   "requests>=2.31.0",
   "polars>0.19.1",
   # Data cleaning dependencies:
-  "dolma[pii,code]@git+https://github.com/allenai/dolma.git@28727302a34225f4ce70ce97940cb18ad9a91583", # Install from git until a 0.9.2 package is released
+  "dolma[pii,code]@git+https://github.com/allenai/dolma.git@e7657e473ec3b46f2c98cd8f61dfab3955ba4755", # Install from git until a 1.0.2 package is released
   "kenlm>=0.2.0", # Used for perplexity tagging
   "blingfire>=0.1.8", # Used for perplexity tagging
 ]
